fix(livereload): patch livereload.js
Livereload errors on ionic dynamic mode links because they don't have an
href attribute.  The project seems to be unmaintained, so it's unlikely
a PR would land any time soon.

The change is:

Reloader.prototype.linkHref = function(link) {
  return link.href || link.getAttribute('data-href');
};

to

Reloader.prototype.linkHref = function(link) {
  return link.href || link.getAttribute('data-href') || "";
};

So that it returns "" and not null, so string operations don't error.