Handle 404 in upload.js
Currently if a 404 is returned you get a json parsing error as seen in driftyco/ionic-cli#1765 which is entirely unhelpful and pretty confusing. This change will show that a 404 was received and give relevant information to whoever gets the error.

Something should probably be done to generally handle any case where the API will return HTML instead of JSON, but this will help alleviate some confusion in the mean time.